src/requirements.txt
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.0.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
colorama==0.4.1
constantly==15.1.0
coverage==4.5.4
cryptography==2.7
daphne==2.3.0
Django==2.2.6
django-multiselectfield==0.1.10
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
more-itertools==7.2.0
msgpack==0.6.2
packaging==19.2
pluggy==0.13.0
postgres==2.2.2
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.19
PyHamcrest==1.9.0
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
pytest-django==3.5.1
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
wcwidth==0.1.7
zipp==0.6.0
zope.interface==4.6.0
